+++ default.nix
@@ -5,4 +5,11 @@ nix-hs {
   cabal = ./byline.cabal;
   flags = [ "build-examples" ];
   compiler = ghc;
+
+  overrides = lib: self: super: {
+    haskeline = if super ? haskeline_0_8_0_0 then
+      lib.dontCheck super.haskeline_0_8_0_0
+    else
+      super.haskeline;
+  };
 }