+++ examples/simple.hs
@@ -1,29 +1,40 @@
-{-# LANGUAGE OverloadedStrings #-}
+-- |
+--
+-- Copyright:
+--   This file is part of the package byline. It is subject to the
+--   license terms in the LICENSE file found in the top-level
+--   directory of this distribution and at:
+--
+--     https://github.com/pjones/byline
+--
+--   No part of this package, including this file, may be copied,
+--   modified, propagated, or distributed except according to the
+--   terms contained in the LICENSE file.
+--
+-- License: BSD-2-Clause
+module Main
+  ( main,
+  )
+where
 
-{-
-
-This file is part of the package byline. It is subject to the license
-terms in the LICENSE file found in the top-level directory of this
-distribution and at git://pmade.com/byline/LICENSE. No part of the
-byline package, including this file, may be copied, modified,
-propagated, or distributed except according to the terms contained in
-the LICENSE file.
-
--}
-
---------------------------------------------------------------------------------
-module Main (main) where
-
---------------------------------------------------------------------------------
-import Control.Monad (void)
-import Data.Text (Text)
+import Byline
+  ( Stylized,
+    askLn,
+    askUntil,
+    bg,
+    blue,
+    bold,
+    fg,
+    red,
+    rgb,
+    runBylineT,
+    sayLn,
+    text,
+  )
 import qualified Data.Text as Text
-import System.Console.Byline
 
---------------------------------------------------------------------------------
 main :: IO ()
-main = void $ runByline $ do
-
+main = void $ runBylineT Nothing $ do
   -- Simple message to stdout:
   sayLn "Okay, let's kick this off"
 
@@ -32,7 +43,7 @@ main = void $ runByline $ do
 
   -- Get user input with a stylized prompt:
   let question = "What's your favorite " <> ("language" <> bold) <> "? "
-  language <- ask question Nothing
+  language <- askLn question Nothing
 
   if Text.null language
     then sayLn "Cat got your tongue?"
@@ -40,14 +51,12 @@ main = void $ runByline $ do
 
   -- Keep prompting until a confirmation function indicates that the
   -- user's input is sufficient:
-  name <- askUntil "What's your name? " Nothing atLeastThreeChars
+  name <- askUntil "What's your name? " Nothing (pure . atLeastThreeChars)
   sayLn $ "Hey there " <> text name <> fg (rgb 108 113 196)
 
---------------------------------------------------------------------------------
 -- | Example confirmation function that requires the input to be three
 -- or more characters long.
-atLeastThreeChars :: Text -> IO (Either Stylized Text)
-atLeastThreeChars input = return $
-  if Text.length input < 3
-    then Left "You can do better."
-    else Right input
+atLeastThreeChars :: Text -> Either (Stylized Text) Text
+atLeastThreeChars input
+  | Text.length input < 3 = Left ("You can do better." <> bg red)
+  | otherwise = Right input